backend/pyproject.toml
[project]
name = "backend"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "django>=5.2.7",
    "django-cors-headers>=4.9.0",
    "django-extensions>=4.1",
    "django-filter>=25.2",
    "django-simpleui",
    "djangorestframework>=3.16.1",
    "djangorestframework-simplejwt>=5.5.1",
    "drf-spectacular>=0.27.0",
    "jionlp>=1.5.27",
    "requests>=2.32.5",
]

[[tool.uv.index]]
name = "aliyun"
url  = "https://mirrors.aliyun.com/pypi/simple/"
